requirements.txt
apache-airflow-providers-snowflake
pandas
requests